feat: add color picker locale and custom popover
